- Added security in UserController. - Added isAdmin field in UserDto.

+++ src/main/java/controllers/UserController.java
@@ -18,22 +18,26 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
+import services.api.UserPermissionService;
 
 import java.beans.IntrospectionException;
 import java.io.IOException;
 import java.lang.reflect.InvocationTargetException;
 
 import static org.springframework.http.HttpStatus.CREATED;
+import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.HttpStatus.OK;
 
 @RestController
 @RequestMapping("/users")
 public class UserController extends AbstractController {
     private final UserFacade userFacade;
+    private final UserPermissionService userPermissionService;
 
     @Autowired
-    public UserController(UserFacade userFacade) {
+    public UserController(UserFacade userFacade, UserPermissionService userPermissionService) {
         this.userFacade = userFacade;
+        this.userPermissionService = userPermissionService;
     }
 
     @PostMapping
@@ -43,19 +47,28 @@ public class UserController extends AbstractController {
     }
 
     @PostMapping("/{userId}")
-    public ResponseEntity<?> updateUserFirstName(@PathVariable Long userId, @RequestBody final String newFirstName) {
+    public ResponseEntity<?> updateUserFirstName(@PathVariable final Long userId, @RequestBody final String newFirstName) {
+        if (!userPermissionService.hasUserAccess(userId)) {
+            return new ResponseEntity<>(FORBIDDEN);
+        }
         userFacade.updateUserFirstName(userId, newFirstName);
         return new ResponseEntity<>(OK);
     }
 
     @GetMapping("/{userId}")
     public ResponseEntity<UserDto> getUserById(@PathVariable final Long userId) {
+        if (!userPermissionService.hasUserAccess(userId) && !(userId == null || userId == 0)) {
+            return new ResponseEntity<>(FORBIDDEN);
+        }
         final UserDto userDto = userFacade.getUserById(userId);
         return new ResponseEntity<>(userDto, OK);
     }
 
     @GetMapping("/{userId}/{informationClass}")
     public ResponseEntity<?> getInformationByClass(final @PathVariable Long userId, @PathVariable final InformationClassDto informationClass) {
+        if (!userPermissionService.hasUserAccess(userId)) {
+            return new ResponseEntity<>(FORBIDDEN);
+        }
         try {
             final InformationDto informationDto = userFacade.getInformationByClass(userId, informationClass);
             return new ResponseEntity<>(informationDto, OK);
@@ -67,6 +80,9 @@ public class UserController extends AbstractController {
 
     @DeleteMapping("/{userId}/{informationClass}")
     public ResponseEntity<?> deleteInformationByClass(final @PathVariable Long userId, @PathVariable final InformationClassDto informationClass) {
+        if (!userPermissionService.hasUserAccess(userId)) {
+            return new ResponseEntity<>(FORBIDDEN);
+        }
         try {
             userFacade.deleteInformationByInformationClass(userId, informationClass);
             return new ResponseEntity<>(OK);
@@ -78,6 +94,9 @@ public class UserController extends AbstractController {
 
     @PostMapping("{userId}/profilePicture")
     public ResponseEntity<UserDto> uploadProfilePicture(@PathVariable Long userId, @RequestParam("file") MultipartFile file) throws IOException {
+        if (!userPermissionService.hasUserAccess(userId)) {
+            return new ResponseEntity<>(FORBIDDEN);
+        }
         if (file.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_ACCEPTABLE);
         }
@@ -87,6 +106,9 @@ public class UserController extends AbstractController {
 
     @DeleteMapping("{userId}/profilePicture")
     public ResponseEntity<UserDto> deleteProfilePicture(@PathVariable Long userId) throws IOException {
+        if (!userPermissionService.hasUserAccess(userId)) {
+            return new ResponseEntity<>(FORBIDDEN);
+        }
         final UserDto userDto = userFacade.deleteProfilePicture(userId);
         return new ResponseEntity<>(userDto, OK);
     }
